refactor: Extracted function (createRegistry)
